Move tests into their respective package tests
