+++ cmd/root/sweet_test.go
@@ -7,9 +7,7 @@ import (
 	"os"
 	"path"
 	"testing"
-	"time"
 
-	"github.com/NicksPatties/sweet/util"
 	"github.com/spf13/cobra"
 )
 
@@ -475,117 +473,4 @@ func TestFromArgsWithEmptyExerciseFiles(t *testing.T) {
 		}
 	}
 
-}
-
-func getEventTs(s string) (t time.Time) {
-	t, _ = time.Parse(util.EventTsLayout, s)
-	return
-}
-
-func TestEventString(t *testing.T) {
-	testCases := []struct {
-		name string
-		in   util.Event
-		want string
-	}{
-		{
-			name: "all fields",
-			in: util.Event{
-				Ts:       getEventTs("2024-10-07 13:46:47.679"),
-				I:        0,
-				Typed:    "a",
-				Expected: "b",
-			},
-			want: "2024-10-07 13:46:47.679\t0\ta\tb",
-		},
-	}
-
-	for _, tc := range testCases {
-		got := fmt.Sprint(tc.in)
-		if got != tc.want {
-			t.Errorf("%s: got\n\t%s\nwant\n\t%s", tc.name, got, tc.want)
-		}
-
-	}
-
-}
-
-func TestParseEvent(t *testing.T) {
-	testCases := []struct {
-		name  string
-		input string
-		want  util.Event
-	}{
-		{
-			name:  "all fields",
-			input: "2024-10-07 13:46:47.679\t0\ta\th",
-			want: util.Event{
-				Ts:       getEventTs("2024-10-07 13:46:47.679"),
-				I:        0,
-				Typed:    "a",
-				Expected: "h",
-			},
-		},
-		{
-			name:  "backspace",
-			input: "2024-10-07 13:46:47.679\t0\tbackspace",
-			want: util.Event{
-				Ts:       getEventTs("2024-10-07 13:46:47.679"),
-				I:        0,
-				Typed:    "backspace",
-				Expected: "",
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		got := util.ParseEvent(tc.input)
-		if !got.Matches(tc.want) {
-			t.Errorf("%s: got\n%s\n\nwant:\n%s", tc.name, got, tc.want)
-		}
-	}
-}
-
-func TestParseEvents(t *testing.T) {
-
-	testCases := []struct {
-		name string
-		in   string
-		want []util.Event
-	}{
-		{
-			name: "two events",
-			in: "2024-10-07 13:46:47.679\t0\ta\th\n" +
-				"2024-10-07 13:46:48.298\t1\tbackspace",
-			want: []util.Event{
-				{
-					Ts:       getEventTs("2024-10-07 13:46:47.679"),
-					I:        0,
-					Typed:    "a",
-					Expected: "h",
-				},
-				{
-					Ts:       getEventTs("2024-10-07 13:46:48.298"),
-					I:        1,
-					Typed:    "backspace",
-					Expected: "",
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		gotEvents := util.ParseEvents(tc.in)
-		for i, got := range gotEvents {
-			if !got.Matches(tc.want[i]) {
-				t.Errorf(
-					"%s [%d]:\ngot\n  %s\nwant\n  %s",
-					tc.name,
-					i,
-					got,
-					tc.want[i],
-				)
-			}
-		}
-	}
 }
